refactor(renderer): scope per-server state to a keyed provider subtree
Everything that belongs to one team server now lives under a `<Show keyed>`
boundary in `app-providers.tsx`, so a server switch discards and rebuilds it
by unmounting. That removes the twenty-setter teardown `selectServer` ran, the
`resetForServer` slice every per-server domain exported for it, and the second
copy of the per-server load sequence that startup and `selectServer` had each
grown their own drifted version of.

The scope key carries a nonce alongside the server id, so a server that is
already active but not yet loaded - a remote that just negotiated a protocol,
or one whose compatibility retry succeeded - remounts instead of being
re-selected over IPC.

Three owners moved above the boundary because their state outlives a switch:
the read-tracking sets (`agent-read-tracking.tsx`), the inline error feed
(`ui-errors.tsx`), and the half of the conversation controller that holds
unfinished user work (`createStableConversationState`). `server-switch.tsx`
holds the three values written by the scope being left and read by the one
being entered.

`activeServerId() !== serverId` after an await is replaced by
`createScopeGuard()`, which asks whether the owner is still alive. The old
comparison got A -> B -> A wrong: the id matches again even though the owner
that started the request was disposed in between. The per-request counters
beside it are kept - they order concurrent writes within one owner, which
disposal cannot express.
