+++ docs/ARCHITECTURE.md
@@ -59,7 +59,16 @@ renderer ──► @openbot/contracts ◄── preload ◄── main ──►
 5. Keep Electron entry points small. New features use a service or a focused IPC input module.
 6. Do not add a second linter or formatter. Biome and its repository-owned anti-slop plugins are the
    only repository lint and format tools.
-7. Do not add temporary compatibility paths without a removal condition and a test for that condition. Released Team API protocol adapters are permanent by default and follow the policy below.
+7. Put renderer state in a domain context module at the root of `src/renderer/src`, and place it by
+   lifetime: state that belongs to one team server goes inside the keyed scope in
+   `app-providers.tsx`, everything else above it. A server switch discards and rebuilds that scope,
+   so it is the only per-server teardown there is - a signal on the wrong side of that boundary
+   either survives a switch it should not or dies in one it should not, and no list of setters can
+   fix it. A context reaches another one with `use*()` only downwards, in the nesting order of
+   `app-providers.tsx`, or through a provider prop; a command that writes to several domains lives
+   in a leaf context or a bridge component mounted under all of them. `window.openbot.*` is not a
+   dependency. Cycles are rejected by `noImportCycles`, so an upward edge must be `import type`.
+8. Do not add temporary compatibility paths without a removal condition and a test for that condition. Released Team API protocol adapters are permanent by default and follow the policy below.
 
 SQLite migration history starts at the frozen version 8 compatibility baseline. Keep the baseline
 schema unchanged, append every later migration in numeric order, and update the separate latest
