+++ src/renderer/src/dynamic-island-bridge.tsx
@@ -8,7 +8,9 @@ import { promptRequestKey } from "./conversation-keys";
 import { useDynamicIsland } from "./dynamic-island";
 import { useNavigation } from "./navigation";
 import { usePlatform } from "./platform";
+import { useServerScope } from "./server-scope";
 import { useServerSelection } from "./server-selection";
+import { useServerSwitch } from "./server-switch";
 import { useServers } from "./servers";
 import { useTurns } from "./turns";
 
@@ -23,21 +25,27 @@ import { useTurns } from "./turns";
  * needs are readable. Splitting them is what keeps the coordinator alive across a
  * server switch while the projection stays scoped to the active one.
  *
- * The projection is withheld until `dynamicIslandLoadedServerId()` matches the
- * active server. Without that, a switch would publish the new server's id next to
- * the previous server's turns for exactly as long as the eight loads take.
+ * The projection is withheld until the scope reports `loaded()`. Without that, a
+ * mount would publish the new server's id next to a half-filled workspace for
+ * exactly as long as its loads take. The flag is a boolean rather than a server
+ * id because the scope it belongs to *is* one server: it starts false on every
+ * mount and cannot describe the wrong one.
  *
  * Actions arrive for *any* server, including one that is not active, which is why
- * `handleDynamicIslandAction` may call `selectServer` before it can act. The two
- * that never need a switch - answering a prompt and responding to an approval -
- * resolve against the coordinator directly and republish, so a reply from the
- * island is reflected before the renderer has caught up.
+ * `handleDynamicIslandAction` may switch before it can act - and why it then
+ * republishes the action through `server-switch.tsx` instead of finishing it: the
+ * switch disposes this bridge, so the scope that lands has to run the rest. The
+ * two that never need a switch - answering a prompt and responding to an
+ * approval - resolve against the coordinator directly and republish, so a reply
+ * from the island is reflected before the renderer has caught up.
  */
 export function DynamicIslandBridge() {
   const platform = usePlatform();
   const { activeServerId } = useServers();
   const { dynamicIslandCoordinator, publishDynamicIslandPresentation } = useDynamicIsland();
-  const { selectServer, dynamicIslandLoadedServerId } = useServerSelection();
+  const { selectServer } = useServerSelection();
+  const { loaded } = useServerScope();
+  const { pendingIslandAction, setPendingIslandAction } = useServerSwitch();
   const { botList, appendUiError } = useAgents();
   const {
     activeTurns,
@@ -52,14 +60,14 @@ export function DynamicIslandBridge() {
   } = useTurns();
   const { unreadReplies, conversationReads, liveMessages } = useConversation();
   const { selectBot, openAgentMessage } = useNavigation();
+  /** The handoff this bridge published, so it never consumes or clears its own. */
+  let publishedAction: DynamicIslandAction | null = null;
 
   createEffect(
     () => {
-      if (platform.landingPreview) return null;
-      const serverId = dynamicIslandLoadedServerId();
-      if (!serverId || serverId !== activeServerId()) return null;
+      if (platform.landingPreview || !loaded()) return null;
       return {
-        serverId,
+        serverId: activeServerId(),
         bots: botList(),
         activeTurns: activeTurns(),
         queues: queues(),
@@ -77,6 +85,29 @@ export function DynamicIslandBridge() {
       if (!input) return;
       dynamicIslandCoordinator.replaceServer(input);
       publishDynamicIslandPresentation();
+    },
+  );
+
+  createEffect(
+    () => ({ action: pendingIslandAction(), ready: loaded() }),
+    ({ action, ready }) => {
+      if (!action) return;
+      // The bridge that published the handoff is still mounted when the signal
+      // changes; consuming or clearing it here would drop the intent before the
+      // switch lands.
+      if (action === publishedAction) return;
+      if (activeServerId() !== action.serverId) {
+        // Another server landed instead - the selection was superseded, or the
+        // user moved on. The intent named a server that is no longer coming.
+        setPendingIslandAction(null);
+        return;
+      }
+      // The action opens a message or a failure inside this workspace, so it has
+      // to wait for the workspace. Acting on a half-loaded scope reads a bot list
+      // and a conversation window that are still empty.
+      if (!ready) return;
+      setPendingIslandAction(null);
+      void handleDynamicIslandAction(action).catch(() => undefined);
     },
   );
 
@@ -117,7 +148,15 @@ export function DynamicIslandBridge() {
       publishDynamicIslandPresentation();
       return;
     }
-    if (activeServerId() !== action.serverId && !(await selectServer(action.serverId, false))) return;
+    if (activeServerId() !== action.serverId) {
+      // The switch replaces this bridge along with the rest of the scope, so the
+      // action is handed to the one that lands rather than finished here against
+      // domains that are about to be disposed.
+      publishedAction = action;
+      setPendingIslandAction(action);
+      if (!(await selectServer(action.serverId, false))) setPendingIslandAction(null);
+      return;
+    }
     selectBot(action.botId);
     if (action.type === "open-message") await openAgentMessage(action.botId, action.messageId);
     if (action.type === "open-failure") {
