+++ src/renderer/src/server-selection.tsx
@@ -1,74 +1,41 @@
 import type { AgentProviderId, BotSummary, ServerSummary } from "@openbot/contracts/ipc";
-import { createEffect, createSignal } from "solid-js";
 import { useAgents } from "./agents";
 import { desktopAnalytics } from "./analytics";
-import { EMPTY_TEAM_PRESENCE } from "./app-defaults";
-import { useBrowserTabs } from "./browser-tabs";
-import { toast } from "./components/ui";
-import { useConversation } from "./conversation";
-import { useDirectMessages } from "./direct-messages";
-import { useDynamicIsland } from "./dynamic-island";
-import { useNavigation } from "./navigation";
-import { usePresence } from "./presence";
 import { useRemoteDesktop } from "./remote-desktop";
+import { useServerSwitch } from "./server-switch";
 import { useServers } from "./servers";
 import { useSettings } from "./settings";
 import { useSetup } from "./setup";
-import { useSidebar } from "./sidebar";
 import { createSimpleContext } from "./simple-context";
-import { useTurns } from "./turns";
 
 /**
  * Switching the active server, and the two ways a new one arrives: joining from
- * an invite, and the servers domain asking for a load it cannot perform itself.
+ * an invite, and opening an agent installed from the marketplace.
  *
- * A leaf, and the largest one. `selectServer` tears down and refills every
- * per-server domain - agents, sidebar, direct messages, conversation, turns,
- * presence, browser - so it can only live below all of them. `servers.tsx` holds
- * the list and the active id but publishes `serverLoadRequest` instead of
- * loading, precisely because the load reaches down here.
+ * `selectServer` used to be the largest leaf in the tree because it tore down
+ * and refilled every per-server domain by hand. It no longer touches any of
+ * them: `setServers` changes the active id, the keyed boundary in
+ * `app-providers.tsx` disposes the scope, and the fresh scope loads itself. What
+ * is left is the part that genuinely spans the two - the compatibility retry,
+ * the analytics, and the authoritative-server recovery when main disagrees with
+ * what was asked for.
  *
- * `dynamicIslandLoadedServerId` is published rather than kept private: it is the
- * flag that says the per-server load finished, and the Dynamic Island projection
- * must not send a half-loaded server to main. It is cleared the moment a switch
- * starts and restored on a failed one, so a failed switch does not leave the
- * island describing a server nobody is looking at.
- *
- * The teardown here is still a list of `resetForServer()` calls. It becomes an
- * unmount once the per-server domains move into a keyed subtree - the point of
- * routing every reset through one exported function per domain was to make that
- * step mechanical.
+ * It still lives inside the scope, below the domains it guards on, and that is
+ * safe for a reason worth naming: every line after `setServers` runs in a
+ * disposed owner, and none of them touch a scoped signal. The browser
+ * suspension and the pending agent selection are read from
+ * `server-switch.tsx`, which sits above the boundary precisely so a value
+ * written before the switch survives it.
  */
 const ServerSelection = createSimpleContext({
   name: "Server selection",
   init: () => {
-    const { servers, setServers, beginServerSelection, serverLoadRequest } = useServers();
+    const { servers, setServers, beginServerSelection } = useServers();
     const { pendingInviteUrl, setPendingInviteUrl, saveSetup } = useSetup();
     const { setSkillsMarketplaceOpen } = useSettings();
     const { disconnectRemoteDesktopWorkspace } = useRemoteDesktop();
-    const { dynamicIslandCoordinator } = useDynamicIsland();
-    const { setTeamPresence } = usePresence();
-    const { cancelDirectConversationRequests, resetForServer: resetDirectMessagesForServer } = useDirectMessages();
-    const {
-      botSetupOpen,
-      creatingAgent,
-      setAgentStatus,
-      setModelOptions,
-      applyStoredBots,
-      resetForServer: resetAgentsForServer,
-    } = useAgents();
-    const { resetForServer: resetTurnsForServer } = useTurns();
-    const {
-      setBrowserVisibilitySuspended,
-      setBrowserControlState,
-      beginBrowserLoad,
-      loadDisplayState: loadBrowserDisplayState,
-      loadControlState: loadBrowserControlState,
-    } = useBrowserTabs();
-    const { setSidebarLayout, loadLayout: loadSidebarLayout, resetForServer: resetSidebarForServer } = useSidebar();
-    const { selectBot } = useNavigation();
-    const { applyConversationReads, resetForServer: resetConversationForServer } = useConversation();
-    const [dynamicIslandLoadedServerId, setDynamicIslandLoadedServerId] = createSignal<string | null>(null);
+    const { setBrowserVisibilitySuspended, setPendingBotSelection } = useServerSwitch();
+    const { botSetupOpen, creatingAgent } = useAgents();
 
     async function selectServer(
       serverId: string,
@@ -88,9 +55,6 @@ const ServerSelection = createSimpleContext({
           await window.openbot.browser.setVisible({ visible: false }).catch(() => undefined);
           if (!selectionIsCurrent()) return false;
         }
-        cancelDirectConversationRequests();
-        const previousDynamicIslandLoadedServerId = dynamicIslandLoadedServerId() ?? previousServerId ?? null;
-        setDynamicIslandLoadedServerId(null);
         let nextServers: ServerSummary[];
         try {
           nextServers = await window.openbot.servers.select(serverId);
@@ -116,7 +80,6 @@ const ServerSelection = createSimpleContext({
               failure_code: "server_select_failed",
             });
           }
-          setDynamicIslandLoadedServerId(previousDynamicIslandLoadedServerId);
           if (recoverAuthoritativeServer) {
             const authoritativeServers = await window.openbot.servers.list().catch(() => null);
             if (!selectionIsCurrent()) return false;
@@ -127,67 +90,19 @@ const ServerSelection = createSimpleContext({
           }
           throw error;
         }
-        const dynamicIslandState = dynamicIslandCoordinator.serverState(serverId);
         setServers(nextServers);
-        resetAgentsForServer();
-        resetSidebarForServer();
-        resetDirectMessagesForServer();
-        resetConversationForServer();
-        resetTurnsForServer(dynamicIslandState);
-        setTeamPresence(EMPTY_TEAM_PRESENCE);
-        const selectedServer = nextServers.find((server) => server.id === serverId);
-        if (
-          selectedServer?.kind === "remote" &&
-          (selectedServer.state === "incompatible" || selectedServer.issue?.code === "protocol_error")
-        ) {
-          return true;
-        }
-        const applyBrowserDisplayState = beginBrowserLoad();
-        const browserDisplayState = loadBrowserDisplayState(selectedServer);
-        const [storedBots, layout, reads, status, models, displayState, controlState, presence] = await Promise.all([
-          window.openbot.agent.listBots(),
-          loadSidebarLayout(selectedServer),
-          window.openbot.agent.listConversationReads(),
-          window.openbot.agent.getStatus(),
-          window.openbot.agent.listModels(),
-          browserDisplayState,
-          loadBrowserControlState(selectedServer),
-          window.openbot.servers.getPresence(),
-        ]);
-        if (!selectionIsCurrent()) return false;
-        setAgentStatus(status);
-        setModelOptions(models);
-        applyBrowserDisplayState(displayState);
-        setBrowserControlState(controlState);
-        setTeamPresence(presence);
-        setSidebarLayout(layout);
-        applyStoredBots(storedBots);
-        applyConversationReads(reads);
-        setDynamicIslandLoadedServerId(serverId);
         return true;
       } finally {
         if (selectionIsCurrent()) setBrowserVisibilitySuspended(false);
       }
     }
 
-    // The servers domain cannot call `selectServer` - it writes to every domain
-    // nested under it - so it publishes the server that needs loading and the load
-    // happens here instead. See `servers.tsx`.
-    createEffect(
-      () => serverLoadRequest(),
-      (request) => {
-        if (!request) return;
-        void selectServer(request.serverId, false).catch((error) => {
-          toast.error("Could not load the remote workspace", {
-            description: error instanceof Error ? error.message : String(error),
-          });
-        });
-      },
-    );
-
     async function openInstalledMarketplaceAgent(bot: BotSummary): Promise<void> {
       if (!(await selectServer("local", false))) return;
-      selectBot(bot.id);
+      // Published rather than called: if this was a switch, the navigation
+      // domain that owns `selectBot` has already been replaced by the one in
+      // the new scope, and that is the one that has to run it.
+      setPendingBotSelection(bot.id);
       setSkillsMarketplaceOpen(false);
     }
 
@@ -223,8 +138,6 @@ const ServerSelection = createSimpleContext({
       joinServer,
       joinRemoteDuringSetup,
       openInstalledMarketplaceAgent,
-      dynamicIslandLoadedServerId,
-      setDynamicIslandLoadedServerId,
     };
   },
 });
