+++ src/renderer/src/setupTests.ts
@@ -3,6 +3,16 @@ import { cleanup } from "@solidjs/testing-library";
 import { afterEach } from "vitest";
 
 export class TestResizeObserver implements ResizeObserver {
+  /**
+   * The observers that are currently watching something.
+   *
+   * Membership follows the observed elements rather than construction, because
+   * the browser keeps an observer alive only while it has a target: one that has
+   * released every element holds nothing open and is collected. Counting
+   * constructed instances instead would report a leak for any library that
+   * releases with `unobserve` rather than `disconnect` - `@tanstack/virtual-core`
+   * is one - and miss nothing in return.
+   */
   static readonly instances = new Set<TestResizeObserver>();
 
   readonly #callback: ResizeObserverCallback;
@@ -10,7 +20,6 @@ export class TestResizeObserver implements ResizeObserver {
 
   constructor(callback: ResizeObserverCallback) {
     this.#callback = callback;
-    TestResizeObserver.instances.add(this);
   }
 
   disconnect(): void {
@@ -20,10 +29,12 @@ export class TestResizeObserver implements ResizeObserver {
 
   observe(element: Element): void {
     this.#elements.add(element);
+    TestResizeObserver.instances.add(this);
   }
 
   unobserve(element: Element): void {
     this.#elements.delete(element);
+    if (this.#elements.size === 0) TestResizeObserver.instances.delete(this);
   }
 
   trigger(element: Element): void {
